主机构的组织信息不能删除

+++ framework-system-module/src/main/java/com/centit/framework/system/controller/UserUnitController.java
@@ -254,6 +254,10 @@ public class UserUnitController extends BaseController {
     public void delete(@PathVariable String userunitid,
                        HttpServletRequest request, HttpServletResponse response) {
         UserUnit dbUserUnit = sysUserUnitManager.getObjectById(userunitid);
+        if("T".equals(dbUserUnit.getIsPrimary())){
+            JsonResultUtils.writeErrorMessageJson("主机构组织信息不能删除！", response);
+            return;
+        }
 
         sysUserUnitManager.deleteObject(dbUserUnit);
